requirements.txt
asn1crypto==1.5.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
configparser==6.0.0
cryptography==44.0.2
et_xmlfile==2.0.0
filelock==3.17.0
greenlet==3.1.1
h5py
idna==3.10
mysql-connector-python==9.2.0
mysql-connector-repackaged==0.3.1
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas
paramiko==3.5.1
platformdirs==4.3.6
pyarrow==18.0.0
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pysmb==1.2.9.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.13.1
six==1.17.0
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sshtunnel==0.4.0
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
schedule